% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceRCT.R
\name{compute_complexity_measures}
\alias{compute_complexity_measures}
\title{Compute complexity measures: entropy, turbulence, volatility}
\usage{
compute_complexity_measures(seqdata)
}
\arguments{
\item{seqdata}{TraMineR sequence object.}
}
\value{
Data frame with Entropy, Turbulence, Volatility for each sequence.
}
\description{
Compute complexity measures: entropy, turbulence, volatility
}
