% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceRCT.R
\name{plot_cluster_membership}
\alias{plot_cluster_membership}
\title{Plot cluster membership distribution by group}
\usage{
plot_cluster_membership(clusters, group)
}
\arguments{
\item{clusters}{Integer vector; output of \code{run_sequence_clustering} or element from \code{analyze_rct_sequences}.}

\item{group}{Factor or character vector; group labels corresponding to clusters.}
}
\value{
ggplot object of cluster frequencies by group.
}
\description{
Plot cluster membership distribution by group
}
