% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceRCT.R
\name{plot_complexity_measures}
\alias{plot_complexity_measures}
\title{Plot complexity measures by group}
\usage{
plot_complexity_measures(
  complexity_df,
  measures = c("Entropy", "Turbulence", "Volatility")
)
}
\arguments{
\item{complexity_df}{Data frame returned by \code{compute_complexity_measures()}, must include \code{Group}.}

\item{measures}{Character vector; complexity metrics to plot (default c("Entropy","Turbulence","Volatility")).}
}
\value{
ggplot object with boxplots and facets for each measure.
}
\description{
Plot complexity measures by group
}
