% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceRCT.R
\name{plot_sequence_index}
\alias{plot_sequence_index}
\title{Plot sequence index by group}
\usage{
plot_sequence_index(seqdata, group, palette = NULL, ...)
}
\arguments{
\item{seqdata}{TraMineR sequence object.}

\item{group}{Group labels.}

\item{palette}{Optional palette.}

\item{...}{Additional args for \code{seqiplot}.}
}
\description{
Plot sequence index by group
}
