% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceRCT.R
\name{plot_state_distribution}
\alias{plot_state_distribution}
\title{Plot state distribution by group}
\usage{
plot_state_distribution(seqdata, group, palette = NULL, ...)
}
\arguments{
\item{seqdata}{TraMineR sequence object created via \code{seqdef}.}

\item{group}{Factor or character vector indicating group for each sequence.}

\item{palette}{Optional color vector for states.}

\item{...}{Additional args passed to \code{seqdplot}.}
}
\description{
Plot state distribution by group
}
