% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceRCT.R
\name{run_sequence_clustering}
\alias{run_sequence_clustering}
\title{Perform hierarchical clustering on sequences}
\usage{
run_sequence_clustering(seqdata, k = 3, method = "ward.D2", distance = "LCS")
}
\arguments{
\item{seqdata}{TraMineR sequence object.}

\item{k}{Integer; number of clusters.}

\item{method}{Character; clustering method (default "ward.D2").}

\item{distance}{Character; seqdist method (default "LCS").}
}
\value{
Integer vector of cluster membership.
}
\description{
Perform hierarchical clustering on sequences
}
